requirements.txt
# Data Science Multi-Agent System Requirements

# Core Framework
crewai>=0.28.0
crewai-tools>=0.2.0
litellm>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning (for reference in modeling agent)
scikit-learn>=1.3.0

# LLM Integration
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Environment Management
python-dotenv>=1.0.0

# Optional: Local LLM Support
# ollama>=0.1.0

# Optional: Alternative LLM Providers
# groq>=0.4.0
# anthropic>=0.8.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0

# For enhanced data analysis (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.14.0
